booking client page details, fix bug ordering, stocks history add and delete quantity

+++ database/migrations/2025_01_10_130909_create_order_items_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('order_id');
             $table->foreignId('product_id');
             $table->decimal('price', 10, 2);
-            $table->decimal('quantity', 10, 2);
+            $table->unsignedInteger('quantity');
             $table->timestamps();
         });
     }
